Fix Abstract2 ordering
Abstract2 ordering (operator <) was defined as the ordering on the abstract
interpretation for the variables alone, not taking the guards into account at
all.

Therefore, two Abstract2 values with the same interpretation for variables but
different interpretations for guards were considered equal, thus losing subsets
sharing the same interpretation for variables when building AbstractSet values.